Generate monitor directory regardless of flag and add director as an issuer to scitoken
