+++ origin_ui/origin.go
@@ -31,6 +31,7 @@ import (
 	"os"
 	"os/signal"
 	"path"
+	"path/filepath"
 	"strings"
 	"sync/atomic"
 	"syscall"
@@ -404,6 +405,37 @@ func ConfigureOriginUI(router *gin.Engine) error {
 	})
 
 	go periodicReload()
+
+	return nil
+}
+
+// Configure XrootD directory for both self-based and director-based file transfer tests
+func ConfigureXrootdMonitoringDir() error {
+	pelicanMonitoringPath := filepath.Join(param.Xrootd_RunLocation.GetString(),
+		"export", "pelican", "monitoring")
+
+	uid, err := config.GetDaemonUID()
+	if err != nil {
+		return err
+	}
+	gid, err := config.GetDaemonGID()
+	if err != nil {
+		return err
+	}
+	username, err := config.GetDaemonUser()
+	if err != nil {
+		return err
+	}
+
+	err = config.MkdirAll(pelicanMonitoringPath, 0755, uid, gid)
+	if err != nil {
+		return errors.Wrapf(err, "Unable to create pelican file trasnfer monitoring directory %v",
+			pelicanMonitoringPath)
+	}
+	if err = os.Chown(pelicanMonitoringPath, uid, -1); err != nil {
+		return errors.Wrapf(err, "Unable to change ownership of pelican file trasnfer monitoring directory %v"+
+			" to desired daemon user %v", pelicanMonitoringPath, username)
+	}
 
 	return nil
 }
