+++ xrootd/authorization.go
@@ -271,6 +271,19 @@ func GenerateOriginIssuer(exportedPaths []string) (issuer Issuer, err error) {
 	return
 }
 
+// We have a special issuer just for self-monitoring the origin.
+func GenerateDirectorMonitoringIssuer() (issuer Issuer, err error) {
+	if val := param.Federation_DirectorUrl.GetString(); val == "" {
+		return
+	}
+	issuer.Name = "Director-based Monitoring"
+	issuer.Issuer = param.Federation_DirectorUrl.GetString()
+	issuer.BasePaths = []string{"/pelican/monitoring"}
+	issuer.DefaultUser = "xrootd"
+
+	return
+}
+
 // Writes out the origin's scitokens.cfg configuration
 func WriteOriginScitokensConfig(exportedPaths []string) error {
 
@@ -307,6 +320,10 @@ func WriteOriginScitokensConfig(exportedPaths []string) error {
 	}
 
 	if issuer, err := GenerateOriginIssuer(exportedPaths); err == nil && len(issuer.Name) > 0 {
+		cfg.Issuers = append(cfg.Issuers, issuer)
+		cfg.Global.Audience = append(cfg.Global.Audience, issuer.Issuer)
+	}
+	if issuer, err := GenerateDirectorMonitoringIssuer(); err == nil && len(issuer.Name) > 0 {
 		cfg.Issuers = append(cfg.Issuers, issuer)
 		cfg.Global.Audience = append(cfg.Global.Audience, issuer.Issuer)
 	}
